Move `sched_*` and `membarrier_*` into `rustix::thread`.
These functions and their associated types operate on threads, rather
than processes.
