+++ src/backend/linux_raw/process/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod cpu_set;
 pub(crate) mod syscalls;
 pub(crate) mod types;
 pub(crate) mod wait;
