+++ src/backend/linux_raw/thread/mod.rs
@@ -1,2 +1,4 @@
+pub(crate) mod cpu_set;
 pub(crate) mod futex;
 pub(crate) mod syscalls;
+pub(crate) mod types;
