+++ src/process/mod.rs
@@ -11,8 +11,6 @@ mod id;
 mod ioctl;
 #[cfg(not(any(target_os = "espidf", target_os = "wasi")))]
 mod kill;
-#[cfg(linux_kernel)]
-mod membarrier;
 #[cfg(target_os = "linux")]
 mod pidfd;
 #[cfg(target_os = "linux")]
@@ -34,9 +32,6 @@ mod procctl;
     target_os = "wasi"
 )))]
 mod rlimit;
-#[cfg(any(freebsdlike, linux_kernel, target_os = "fuchsia"))]
-mod sched;
-mod sched_yield;
 #[cfg(not(target_os = "wasi"))] // WASI doesn't have umask.
 mod umask;
 #[cfg(not(any(target_os = "espidf", target_os = "vita", target_os = "wasi")))]
@@ -53,8 +48,6 @@ pub use id::*;
 pub use ioctl::*;
 #[cfg(not(any(target_os = "espidf", target_os = "wasi")))]
 pub use kill::*;
-#[cfg(linux_kernel)]
-pub use membarrier::*;
 #[cfg(target_os = "linux")]
 pub use pidfd::*;
 #[cfg(target_os = "linux")]
@@ -75,9 +68,6 @@ pub use procctl::*;
     target_os = "wasi"
 )))]
 pub use rlimit::*;
-#[cfg(any(freebsdlike, linux_kernel, target_os = "fuchsia"))]
-pub use sched::*;
-pub use sched_yield::sched_yield;
 #[cfg(not(target_os = "wasi"))]
 pub use umask::*;
 #[cfg(not(any(target_os = "espidf", target_os = "vita", target_os = "wasi")))]
