+++ src/thread/mod.rs
@@ -9,7 +9,12 @@ mod id;
 #[cfg(linux_kernel)]
 mod libcap;
 #[cfg(linux_kernel)]
+mod membarrier;
+#[cfg(linux_kernel)]
 mod prctl;
+#[cfg(any(freebsdlike, linux_kernel, target_os = "fuchsia"))]
+mod sched;
+mod sched_yield;
 #[cfg(linux_kernel)]
 mod setns;
 
@@ -20,13 +25,15 @@ pub use crate::thread::futex::{
 #[cfg(not(target_os = "redox"))]
 pub use clock::*;
 #[cfg(linux_kernel)]
-pub use id::{
-    gettid, set_thread_gid, set_thread_groups, set_thread_res_gid, set_thread_res_uid,
-    set_thread_uid, Gid, Pid, RawGid, RawPid, RawUid, Uid,
-};
+pub use id::*;
 #[cfg(linux_kernel)]
 pub use libcap::{capabilities, set_capabilities, CapabilityFlags, CapabilitySets};
 #[cfg(linux_kernel)]
+pub use membarrier::*;
+#[cfg(linux_kernel)]
 pub use prctl::*;
+#[cfg(any(freebsdlike, linux_kernel, target_os = "fuchsia"))]
+pub use sched::*;
+pub use sched_yield::sched_yield;
 #[cfg(linux_kernel)]
 pub use setns::*;
