+++ src/thread/sched.rs
@@ -1,4 +1,4 @@
-use crate::process::Pid;
+use crate::pid::Pid;
 use crate::{backend, io};
 use core::{fmt, hash};
 
@@ -11,23 +11,23 @@ use core::{fmt, hash};
 ///  - [Linux]
 ///
 /// [Linux]: https://man7.org/linux/man-pages/man3/CPU_SET.3.html
-/// [`sched_setaffinity`]: crate::process::sched_setaffinity
-/// [`sched_getaffinity`]: crate::process::sched_getaffinity
+/// [`sched_setaffinity`]: crate::thread::sched_setaffinity
+/// [`sched_getaffinity`]: crate::thread::sched_getaffinity
 #[repr(C)]
 #[derive(Clone, Copy)]
 pub struct CpuSet {
-    cpu_set: backend::process::types::RawCpuSet,
+    cpu_set: backend::thread::types::RawCpuSet,
 }
 
 impl CpuSet {
     /// The maximum number of CPU in `CpuSet`.
-    pub const MAX_CPU: usize = backend::process::types::CPU_SETSIZE;
+    pub const MAX_CPU: usize = backend::thread::types::CPU_SETSIZE;
 
     /// Create a new and empty `CpuSet`.
     #[inline]
     pub fn new() -> Self {
         Self {
-            cpu_set: backend::process::types::raw_cpu_set_new(),
+            cpu_set: backend::thread::types::raw_cpu_set_new(),
         }
     }
 
@@ -36,7 +36,7 @@ impl CpuSet {
     /// `field` is the CPU id to test.
     #[inline]
     pub fn is_set(&self, field: usize) -> bool {
-        backend::process::cpu_set::CPU_ISSET(field, &self.cpu_set)
+        backend::thread::cpu_set::CPU_ISSET(field, &self.cpu_set)
     }
 
     /// Add a CPU to `CpuSet`.
@@ -44,7 +44,7 @@ impl CpuSet {
     /// `field` is the CPU id to add.
     #[inline]
     pub fn set(&mut self, field: usize) {
-        backend::process::cpu_set::CPU_SET(field, &mut self.cpu_set)
+        backend::thread::cpu_set::CPU_SET(field, &mut self.cpu_set)
     }
 
     /// Remove a CPU from `CpuSet`.
@@ -52,20 +52,20 @@ impl CpuSet {
     /// `field` is the CPU id to remove.
     #[inline]
     pub fn unset(&mut self, field: usize) {
-        backend::process::cpu_set::CPU_CLR(field, &mut self.cpu_set)
+        backend::thread::cpu_set::CPU_CLR(field, &mut self.cpu_set)
     }
 
     /// Count the number of CPUs set in the `CpuSet`.
     #[cfg(linux_kernel)]
     #[inline]
     pub fn count(&self) -> u32 {
-        backend::process::cpu_set::CPU_COUNT(&self.cpu_set)
+        backend::thread::cpu_set::CPU_COUNT(&self.cpu_set)
     }
 
     /// Zeroes the `CpuSet`.
     #[inline]
     pub fn clear(&mut self) {
-        backend::process::cpu_set::CPU_ZERO(&mut self.cpu_set)
+        backend::thread::cpu_set::CPU_ZERO(&mut self.cpu_set)
     }
 }
 
@@ -107,7 +107,7 @@ impl Eq for CpuSet {}
 
 impl PartialEq for CpuSet {
     fn eq(&self, other: &Self) -> bool {
-        backend::process::cpu_set::CPU_EQUAL(&self.cpu_set, &other.cpu_set)
+        backend::thread::cpu_set::CPU_EQUAL(&self.cpu_set, &other.cpu_set)
     }
 }
 
@@ -125,7 +125,7 @@ impl PartialEq for CpuSet {
 /// [Linux]: https://man7.org/linux/man-pages/man2/sched_setaffinity.2.html
 #[inline]
 pub fn sched_setaffinity(pid: Option<Pid>, cpuset: &CpuSet) -> io::Result<()> {
-    backend::process::syscalls::sched_setaffinity(pid, &cpuset.cpu_set)
+    backend::thread::syscalls::sched_setaffinity(pid, &cpuset.cpu_set)
 }
 
 /// `sched_getaffinity(pid)`—Get a thread's CPU affinity mask.
@@ -142,7 +142,7 @@ pub fn sched_setaffinity(pid: Option<Pid>, cpuset: &CpuSet) -> io::Result<()> {
 #[inline]
 pub fn sched_getaffinity(pid: Option<Pid>) -> io::Result<CpuSet> {
     let mut cpuset = CpuSet::new();
-    backend::process::syscalls::sched_getaffinity(pid, &mut cpuset.cpu_set).and(Ok(cpuset))
+    backend::thread::syscalls::sched_getaffinity(pid, &mut cpuset.cpu_set).and(Ok(cpuset))
 }
 
 /// `sched_getcpu()`—Get the CPU that the current thread is currently on.
@@ -157,5 +157,5 @@ pub fn sched_getaffinity(pid: Option<Pid>) -> io::Result<CpuSet> {
 #[cfg(any(linux_kernel, target_os = "dragonfly"))]
 #[inline]
 pub fn sched_getcpu() -> usize {
-    backend::process::syscalls::sched_getcpu()
+    backend::thread::syscalls::sched_getcpu()
 }
