+++ tests/process/main.rs
@@ -4,11 +4,8 @@
 #![cfg(not(windows))]
 #![cfg_attr(core_c_str, feature(core_c_str))]
 
-mod cpu_set;
 #[cfg(not(target_os = "wasi"))] // WASI doesn't have get[gpu]id.
 mod id;
-#[cfg(linux_kernel)]
-mod membarrier;
 #[cfg(target_os = "linux")]
 mod pidfd;
 #[cfg(linux_kernel)]
@@ -19,7 +16,6 @@ mod priority;
 mod procctl;
 #[cfg(not(any(target_os = "fuchsia", target_os = "redox", target_os = "wasi")))]
 mod rlimit;
-mod sched;
 #[cfg(not(target_os = "wasi"))] // WASI doesn't have umask.
 mod umask;
 #[cfg(not(any(target_os = "espidf", target_os = "wasi")))] // WASI doesn't have waitpid.
