+++ tests/thread/cpu_set.rs
@@ -1,10 +1,10 @@
 #[cfg(linux_kernel)]
 #[test]
 fn test_cpu_set() {
-    let set = rustix::process::sched_getaffinity(None).unwrap();
+    let set = rustix::thread::sched_getaffinity(None).unwrap();
 
     let mut count = 0;
-    for i in 0..rustix::process::CpuSet::MAX_CPU {
+    for i in 0..rustix::thread::CpuSet::MAX_CPU {
         if set.is_set(i) {
             count += 1;
         }
