+++ tests/thread/main.rs
@@ -5,6 +5,7 @@
 
 #[cfg(not(target_os = "redox"))]
 mod clocks;
+mod cpu_set;
 #[cfg(linux_kernel)]
 mod futex;
 #[cfg(linux_kernel)]
@@ -12,6 +13,9 @@ mod id;
 #[cfg(linux_kernel)]
 mod libcap;
 #[cfg(linux_kernel)]
+mod membarrier;
+#[cfg(linux_kernel)]
 mod prctl;
+mod sched;
 #[cfg(linux_kernel)]
 mod setns;
